chore(example): use 'runtime' instead of building a package.json file
In order to define which Node.js version the Cloud Function should use,
we were copying the `package.json` file and adding an `engines.node`
entry to it. Now we'll use the original `package.json` file and just
specify the Node.js version with the new `functions.runtime` entry in
`firebase.json`.